services/api/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
sentence-transformers
numpy
